+++ cmd/review/config.go
@@ -1,0 +1,10 @@
+package main
+
+var opts struct {
+	GithubToken string `long:"gh-token" env:"GITHUB_TOKEN" description:"GitHub token" required:"true"`
+	OpenAIToken string `long:"openai-token" env:"OPENAI_TOKEN" description:"OpenAI token" required:"true"`
+	Owner       string `long:"owner" env:"OWNER" description:"GitHub owner" required:"true"`
+	Repo        string `long:"repo" env:"REPO" description:"GitHub repo" required:"true"`
+	PRNumber    int    `long:"pr-number" env:"PR_NUMBER" description:"Pull request number" required:"true"`
+	Test        bool   `long:"test" env:"TEST" description:"Test mode"`
+}
