+++ go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/google/go-github/v51 v51.0.0
 	github.com/jessevdk/go-flags v1.5.0
 	github.com/sashabaranov/go-openai v1.7.0
+	golang.org/x/oauth2 v0.6.0
 )
 
 require (
@@ -15,7 +16,6 @@ require (
 	github.com/google/go-querystring v1.1.0 // indirect
 	golang.org/x/crypto v0.7.0 // indirect
 	golang.org/x/net v0.8.0 // indirect
-	golang.org/x/oauth2 v0.6.0 // indirect
 	golang.org/x/sys v0.6.0 // indirect
 	google.golang.org/appengine v1.6.7 // indirect
 	google.golang.org/protobuf v1.28.0 // indirect
